Updates to AutoMapper 5.0
After 3.x there was a significant move to an instance based configuration
rather than a global configuration. 1.x of this library fails against the
latest version. This adds the ability to pass the current configuration in
as part of the query.

Because there are extra parameters I bumped this to a 2.0 release to
indicate the breaking change.